+++ src/Maps/utils/namespace.ts
@@ -154,17 +154,4 @@ export namespace MapUtils {
         readonly mapboxAttr: string;
         readonly hereMapsAttr: string;
     }
-
-    export const customUrls: CustomTypeUrls = {
-        openStreetMap: `https://{s}.tile.osm.org/{z}/{x}/{y}.png`,
-        mapbox: `https://api.tiles.mapbox.com/v4/{id}/{z}/{x}/{y}.png?access_token=`,
-        hereMaps: `https://2.base.maps.cit.api.here.com/maptile/2.1/maptile/newest/normal.day/{z}/{x}/{y}/256/png8?`
-    };
-
-    export const mapAttr: MapAttributions = {
-        openStreetMapAttr: `&copy; <a href="https://osm.org/copyright">OpenStreetMap</a> contributors`,
-        mapboxAttr: `Map data &copy; <a href="https://www.openstreetmap.org/">OpenStreetMap</a> contributors,
-            <a href="https://creativecommons.org/licenses/by-sa/2.0/">CC-BY-SA</a>, Imagery © <a href="https://www.mapbox.com/">Mapbox</a>`,
-        hereMapsAttr: `Map &copy; 1987-2014 <a href="https://developer.here.com">HERE</a>`
-    };
 }
