--- src/Maps/utils/sharedConfigs.ts
@@ -1,22 +1,0 @@
-import { CSSProperties } from "react";
-import { MapUtils } from "./namespace";
-import Dimensions = MapUtils.Dimensions;
-
-export namespace Shared {
-    export const getDimensions = <T extends Dimensions>(props: T): CSSProperties => {
-        const style: CSSProperties = {
-            width: props.widthUnit === "percentage" ? `${props.width}%` : `${props.width}px`
-        };
-        if (props.heightUnit === "percentageOfWidth") {
-            style.paddingBottom = props.widthUnit === "percentage"
-                ? `${props.height}%`
-                : `${props.width / 2}px`;
-        } else if (props.heightUnit === "pixels") {
-            style.height = `${props.height}px`;
-        } else if (props.heightUnit === "percentageOfParent") {
-            style.height = `${props.height}%`;
-        }
-
-        return style;
-    };
-}
